[PORT] base_delivery_carrier_label port to 9.0 - Reactivate - Move files in views and models - Adapt XML views - Update README.rst
Move settings menu in Inventory/Configuration

Remove test with Logistic units as it was removed from 9.0

Allow to give a filename to the created label if not defined use the attachment name

Fix weight calcul adding a field for total weight
This due to removed logistics units
If set this field will be used to define weight, otherwise we compute
weight even before quants have been created.

Fix stock picking view some labels were missing and the view was scrambled

Split files by model
