+++ base_delivery_carrier_label/models/__init__.py
@@ -1,0 +1,8 @@
+from . import delivery_carrier
+from . import delivery_carrier_option
+from . import delivery_carrier_template_option
+from . import stock_pack_operation
+from . import stock_picking
+from . import stock_quant_package
+from . import shipping_label
+from . import carrier_account
