--- base_delivery_carrier_label/stock_view.xml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-  <data>
-
-    <!-- From DELIVERY module on STOCK.PICKING -->
-    <!-- FORM -->
-    <record id="view_picking_withcarrier_out_form" model="ir.ui.view">
-      <field name="model">stock.picking</field>
-      <field name="inherit_id" ref="delivery.view_picking_withcarrier_out_form"/>
-      <field name="arch" type="xml">
-        <field name="state" position="before">
-          <button name="action_generate_carrier_label" states="done"
-            string="Create Shipping Label" type="object"/>
-        </field>
-        <xpath expr="//page[@string='Additional Info']//field[@name='carrier_tracking_ref']" position="after">
-          <field name="carrier_type"/>
-          <field name="carrier_code"/>
-          <field name="option_ids" nolabel="1" colspan="2"
-            domain="[('carrier_id', '=', carrier_id)]"/>
-        </xpath>
-      </field>
-    </record>
-
-
-
-<record id="vpicktree" model="ir.ui.view">
-    <field name="model">stock.picking</field>
-    <field name="inherit_id" ref="stock.vpicktree"/>
-    <field name="arch" type="xml">
-        <field name="state" position="before">
-            <field name="carrier_id"/>
-        </field>
-    </field>
-</record>
-
-<record id="view_quant_package_form" model="ir.ui.view">
-    <field name="model">stock.quant.package</field>
-    <field name="inherit_id" ref="stock.view_quant_package_form"/>
-    <field name="arch" type="xml">
-        <field name="ul_id" position="before">
-            <field name="parcel_tracking"/>
-            <field name="weight"/>
-        </field>
-    </field>
-</record>
-
-<record id="view_quant_package_tree" model="ir.ui.view">
-    <field name="model">stock.quant.package</field>
-    <field name="inherit_id" ref="stock.view_quant_package_tree"/>
-    <field name="arch" type="xml">
-        <field name="complete_name" position="after">
-            <field name="parcel_tracking"/>
-            <field name="weight"/>
-        </field>
-    </field>
-</record>
-
-<record id="quant_package_search_view" model="ir.ui.view">
-    <field name="model">stock.quant.package</field>
-    <field name="inherit_id" ref="stock.quant_package_search_view"/>
-    <field name="arch" type="xml">
-        <field name="name" position="after">
-            <field name="parcel_tracking"/>
-        </field>
-    </field>
-</record>
-
-
-</data>
-</openerp>
