+++ base_delivery_carrier_label/tests/test_get_weight.py
@@ -13,13 +13,11 @@ class TestGetWeight(TransactionCase):
         })
 
     def _create_order_line(self, order, products):
-        ol = []
         for product in products:
-            ol.append(self.env['sale.order.line'].create({
+            self.env['sale.order.line'].create({
                 'product_id': product.id,
                 'order_id': order.id,
-            }))
-        return ol
+            })
 
     def _create_ul(self):
         vals = [{
@@ -73,26 +71,27 @@ class TestGetWeight(TransactionCase):
         customer = self.env['res.partner'].search([], limit=1)
         order = self._create_order(customer)
         self._create_order_line(order, products)
-        order.action_button_confirm()
+        order.action_confirm()
         picking = order.picking_ids
         picking.do_transfer()
         return picking
 
     def test_get_weight(self):
-        """Test quant.package.get_weight and pack.operation.get_weight."""
+        """Test quant.package.weight computed field and
+        pack.operation.get_weight."""
         # prepare some data
         weights = [2, 30, 1, 24, 39]
         products = self._get_products(weights)
         picking = self._generate_picking(products)
         package = self.env['stock.quant.package'].create({})
-        operations = []
+        operations = self.env['stock.pack.operation']
         for product in products:
-            operations.append(self._create_operation(picking, {
+            operations |= self._create_operation(picking, {
                 'product_qty': 1,
                 'product_id': product.id,
                 'product_uom_id': product.uom_id.id,
                 'result_package_id': package.id,
-            }))
+            })
         # end of prepare data
 
         # test operation.get_weight()
@@ -101,10 +100,40 @@ class TestGetWeight(TransactionCase):
                 operation.get_weight(),
                 operation.product_id.weight * operation.product_qty)
 
-        # test package.get_weight()
+        # test package.weight
         self.assertEqual(
-            package.get_weight(),
+            package.weight,
             sum([product.weight for product in products]))
+
+    def test_total_weight(self):
+        """Test quant.package.weight computed field when a total
+        weight is defined """
+        # prepare some data
+        weights = [2, 30, 1, 24, 39]
+        products = self._get_products(weights)
+        picking = self._generate_picking(products)
+        package = self.env['stock.quant.package'].create({})
+        operations = self.env['stock.pack.operation']
+        for product in products:
+            operations |= self._create_operation(picking, {
+                'product_qty': 1,
+                'product_id': product.id,
+                'product_uom_id': product.uom_id.id,
+                'result_package_id': package.id,
+            })
+        package.total_weight = 1542.0
+        # end of prepare data
+
+        # test operation.get_weight()
+        for operation in operations:
+            self.assertEqual(
+                operation.get_weight(),
+                operation.product_id.weight * operation.product_qty)
+
+        # test package.weight
+        self.assertEqual(
+            package.weight,
+            package.total_weight)
 
     def test_get_weight_with_qty(self):
         """Ensure qty are taken in account."""
@@ -113,14 +142,14 @@ class TestGetWeight(TransactionCase):
         products = self._get_products(weights)
         picking = self._generate_picking(products)
         package = self.env['stock.quant.package'].create({})
-        operations = []
+        operations = self.env['stock.pack.operation']
         for idx, product in enumerate(products):
-            operations.append(self._create_operation(picking, {
+            operations |= self._create_operation(picking, {
                 'product_qty': idx,  # nice one
                 'product_id': product.id,
                 'product_uom_id': product.uom_id.id,
                 'result_package_id': package.id
-            }))
+            })
         # end of prepare data
 
         # test operation.get_weight()
@@ -129,51 +158,10 @@ class TestGetWeight(TransactionCase):
                 operation.get_weight(),
                 operation.product_id.weight * operation.product_qty)
 
-        # test package.get_weight()
+        # test package._weight
         self.assertEqual(
-            package.get_weight(),
+            package.weight,
             sum([operation.get_weight() for operation in operations]))
-
-    def test_get_weight_with_lu(self):
-        """Check with logistic unit."""
-        # prepare some data
-        weights = [2.39, 3.1, 20, 24, 39]
-        products = self._get_products(weights)
-        picking = self._generate_picking(products)
-        uls = self._create_ul()
-        packages = [
-            self.env['stock.quant.package'].create(
-                {'ul_id': ul.id}
-            ) for ul in uls
-        ]
-        operations = []
-        for idx, product in enumerate(products):
-            pack = packages[idx % len(packages)]
-            operations.append(self._create_operation(picking, {
-                'product_qty': 1,
-                'product_id': product.id,
-                'product_uom_id': product.uom_id.id,
-                'result_package_id': pack.id,
-            }))
-        # end of prepare data
-
-        products_weight = sum([product.weight for product in products])
-        packages_weight = sum(
-            [package.ul_id.weight for package in packages])
-
-        self.assertEqual(
-            sum([package.get_weight() for package in packages]),
-            packages_weight + products_weight)
-
-        # with @api.multi
-        packages_multi = self.env['stock.quant.package'].search(
-            [['id', 'in', [package.id for package in packages]]]
-        )
-
-        self.assertEqual(
-            packages_multi.get_weight(),
-            packages_weight + products_weight
-        )
 
     def test_get_weight_with_uom(self):
         """Check with differents uom."""
@@ -188,7 +176,6 @@ class TestGetWeight(TransactionCase):
             {
                 'name': 'Expected Odoo dev documentation',
                 'uom_id': tonne_id.id,
-                'uos_id': tonne_id.id,
                 'uom_po_id': tonne_id.id,
                 'weight': weights[0]
             })
@@ -197,7 +184,6 @@ class TestGetWeight(TransactionCase):
             {
                 'name': 'OCA documentation',
                 'uom_id': kg_id.id,
-                'uos_id': kg_id.id,
                 'uom_po_id': kg_id.id,
                 'weight': weights[1],
             }))
@@ -205,7 +191,6 @@ class TestGetWeight(TransactionCase):
             {
                 'name': 'Actual Odoo dev documentation',
                 'uom_id': gr_id.id,
-                'uos_id': gr_id.id,
                 'uom_po_id': gr_id.id,
                 'weight': weights[2],
             })
@@ -216,18 +201,18 @@ class TestGetWeight(TransactionCase):
             weights[2] * 0.01  # g
         )
         picking = self._generate_picking(products)
-        operations = []
+        operations = self.env['stock.pack.operation']
         for product in products:
-            operations.append(self._create_operation(picking, {
+            operations |= self._create_operation(picking, {
                 'product_qty': 1,
                 'product_id': product.id,
                 'product_uom_id': product.uom_id.id,
                 'result_package_id': package.id,
-            }))
+            })
         # end of prepare data
 
         # because uom conversion is not implemented
-        self.assertEqual(package.get_weight(), False)
+        self.assertEqual(package.weight, False)
 
         # if one day, uom conversion is implemented:
         # self.assertEqual(package.get_weight(), products_weight)
