+++ base_delivery_carrier_label/views/delivery.xml
@@ -1,0 +1,102 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+
+  <!-- VIEW FOR THE OBJECT : delivery_carrier_template_option -->
+  <record id="delivery_carrier_template_option_view_form" model="ir.ui.view">
+    <field name="name">delivery_base.delivery_carrier_option.view_form</field>
+    <field name="model">delivery.carrier.template.option</field>
+    <field name="arch" type="xml">
+      <form string="delivery_carrier_option">
+        <field name="partner_id" />
+        <field name="code"/>
+        <field name="name"/>
+        <newline/>
+        <field name="description" colspan="4"/>
+      </form>
+    </field>
+  </record>
+
+  <record id="delivery_carrier_template_option_view_tree" model="ir.ui.view">
+    <field name="name">delivery_base.delivery_carrier_template_option.view_tree</field>
+    <field name="model">delivery.carrier.template.option</field>
+    <field name="arch" type="xml">
+      <tree string="delivery_carrier_option">
+        <field name="partner_id" />
+        <field name="code"/>
+        <field name="name"/>
+      </tree>
+    </field>
+  </record>
+
+  <!-- VIEW FOR THE OBJECT : delivery_carrier_option -->
+  <record id="delivery_carrier_option_view_form" model="ir.ui.view">
+    <field name="name">delivery_base.delivery_carrier_option.view_form</field>
+    <field name="model">delivery.carrier.option</field>
+    <field name="arch" type="xml">
+      <form string="delivery_carrier_option">
+        <group col="4">
+          <field name="mandatory" attrs="{'readonly': [('readonly_flag','=',True)]}"/>
+          <field name="by_default" attrs="{'invisible': [('mandatory', '=', True)]}"/>
+          <field name="readonly_flag" attrs="{'invisible': True}"/>
+          <field name="tmpl_option_id" colspan="4"
+            attrs="{'readonly': [('readonly_flag','=',True)]}"/>
+          <newline/>
+          <field name="description" colspan="4" readonly="True"/>
+        </group>
+      </form>
+    </field>
+  </record>
+
+  <record id="delivery_carrier_option_view_tree" model="ir.ui.view">
+    <field name="name">delivery_base.delivery_carrier_option.view_tree</field>
+    <field name="model">delivery.carrier.option</field>
+    <field name="type">tree</field>
+    <field name="arch" type="xml">
+      <tree string="delivery_carrier_option">
+        <field name="mandatory" />
+        <field name="tmpl_option_id" />
+        <field name="code"/>
+      </tree>
+    </field>
+  </record>
+
+  <!-- INHERITED VIEW FOR THE OBJECT : delivery.carrier -->
+
+  <record id="view_delivery_carrier_form" model="ir.ui.view">
+    <field name="name">delivery_base.delivery.carrier.view_form</field>
+    <field name="model">delivery.carrier</field>
+    <field name="inherit_id" ref="delivery.view_delivery_carrier_form"/>
+    <field name="arch" type="xml">
+
+      <xpath expr="//h1" position="after">
+        <group>
+          <field name="type"/>
+          <field name="code"/>
+        </group>
+      </xpath>
+
+      <xpath expr="//notebook" position="inside">
+        <page string="Options">
+          <field name="available_option_ids" nolabel="1" colspan="4"/>
+        </page>
+        <page string="Description">
+          <field name="description" colspan="4" nolabel="1"/>
+        </page>
+      </xpath>
+
+    </field>
+  </record>
+
+  <record id="view_delivery_carrier_tree" model="ir.ui.view">
+    <field name="name">delivery_carrier_base.tree</field>
+    <field name="model">delivery.carrier</field>
+    <field name="inherit_id" ref="delivery.view_delivery_carrier_tree"/>
+    <field name="arch" type="xml">
+      <field name="name" position="after">
+        <field name="type"/>
+        <field name="code"/>
+      </field>
+    </field>
+  </record>
+
+</odoo>
