+++ base_delivery_carrier_label/views/stock.xml
@@ -1,0 +1,67 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+
+  <!-- From DELIVERY module on STOCK.PICKING -->
+  <!-- FORM -->
+  <record id="view_picking_withcarrier_out_form" model="ir.ui.view">
+    <field name="model">stock.picking</field>
+    <field name="inherit_id" ref="delivery.view_picking_withcarrier_out_form"/>
+    <field name="arch" type="xml">
+      <field name="state" position="before">
+        <button name="action_generate_carrier_label" states="done"
+          string="Create Shipping Label" type="object"/>
+      </field>
+      <xpath expr="//page//field[@name='carrier_id']" position="after">
+        <field name="carrier_type"/>
+        <field name="carrier_code"/>
+      </xpath>
+      <xpath expr="//page//group[@name='carrier_data']/.." position="after">
+        <field name="option_ids"
+          domain="[('carrier_id', '=', carrier_id)]"/>
+      </xpath>
+    </field>
+  </record>
+
+  <record id="vpicktree" model="ir.ui.view">
+    <field name="model">stock.picking</field>
+    <field name="inherit_id" ref="stock.vpicktree"/>
+    <field name="arch" type="xml">
+      <field name="state" position="before">
+        <field name="carrier_id"/>
+      </field>
+    </field>
+  </record>
+
+  <record id="view_quant_package_form" model="ir.ui.view">
+    <field name="model">stock.quant.package</field>
+    <field name="inherit_id" ref="stock.view_quant_package_form"/>
+    <field name="arch" type="xml">
+      <field name="packaging_id" position="before">
+        <field name="parcel_tracking"/>
+        <field name="weight"/>
+      </field>
+    </field>
+  </record>
+
+  <record id="view_quant_package_tree" model="ir.ui.view">
+    <field name="model">stock.quant.package</field>
+    <field name="inherit_id" ref="stock.view_quant_package_tree"/>
+    <field name="arch" type="xml">
+      <field name="complete_name" position="after">
+        <field name="parcel_tracking"/>
+        <field name="weight"/>
+      </field>
+    </field>
+  </record>
+
+  <record id="quant_package_search_view" model="ir.ui.view">
+    <field name="model">stock.quant.package</field>
+    <field name="inherit_id" ref="stock.quant_package_search_view"/>
+    <field name="arch" type="xml">
+      <field name="name" position="after">
+        <field name="parcel_tracking"/>
+      </field>
+    </field>
+  </record>
+
+</odoo>
